change some fn names and move some code into dedicated files
